Feature/26/implementacion nombrecurso en carrito (#27)
* agregando el parametro nombre curso en carrito

* correcion envio de mensaje de error de mercado pago

* control de errores en use cases

+++ src/application/uses-cases/Carrito-recovery/delete-carrito.usecase.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/core/services/prisma/prisma.service';
 
 @Injectable()
@@ -6,6 +6,14 @@ export class EliminarCarritoUseCase {
   constructor(private readonly prisma: PrismaService) {}
 
   async execute(id: number) {
+    const carritoExistente = await this.prisma.carrito.findUnique({
+      where: { id },
+    });
+
+    if (!carritoExistente) {
+      throw new NotFoundException(`Carrito con id ${id} no encontrado`);
+    }
+
     await this.prisma.carrito.delete({
       where: { id },
     });
